Group some components in folders
