+++ lib/shared/App/HotkeyedMenuItem.jsx
