+++ lib/shared/App/Splitter.jsx
@@ -35,13 +35,25 @@
  */
 
 import React from 'react';
-import spinnerImg from '../../resources/ajax-loader.gif';
 
-export default () => (
-    <img
-        src={spinnerImg}
-        height="16"
-        width="16"
-        alt="Loading..."
+import '../../../resources/css/brand19/splitter.scss';
+
+const onMouseDownHorizontal = () => console.log('FIXME onMouseDownHorizontal');
+export const HorizontalSplitter = () => (
+    <div
+        tabIndex={-1}
+        role="button"
+        className="core19-splitter horizontal"
+        onMouseDown={event => onMouseDownHorizontal(event/* , resizeLogContainer */)}
+    />
+);
+
+const onMouseDownVertical = () => console.log('FIXME onMouseDownVertical');
+export const VerticalSplitter = () => (
+    <div
+        tabIndex={-1}
+        role="button"
+        className="core19-splitter vertical"
+        onMouseDown={onMouseDownVertical}
     />
 );
