+++ lib/shared/Dialog/ConfirmationDialog.jsx
