+++ lib/shared/index.js
@@ -34,8 +34,8 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export { default as App } from './App';
-export { default as NavBar } from './NavBar';
+export { default as App } from './App/App';
+export { default as NavBar } from './App/NavBar';
 export { default as DeviceSelector } from './Device/DeviceSelector';
 
 export { default as coreReducer } from './coreReducer';
